Update to PythonV3 and Fixed subsequent errors
Ran the 2to3.py updater on the original file,
Removed the redundant line of urllib imports that resulted from having both urllib and urllib2 in the list,
Added .decode() to the response.read() call because of a type error between 'bytes' and 'string',
Removed the '\\'+ prefix from the query at line 145 so that the actual query is passed to the API without the pollution of a \ 
ie. ('1+1') instead of ('\1+1'),
Gave it a module name.